chore: migrate trello to package
