+++ docs/_snippets/apps.mdx
@@ -332,6 +332,15 @@
       </Card>
     </a>
       
+    <a href="https://activepieces.com/docs/pieces/apps/trello">
+      <Card>
+          <p align="center">
+            <strong>trello</strong>
+            <img height="75px" width="75px" src="https://cdn.activepieces.com/pieces/trello.png" />
+          </p>
+      </Card>
+    </a>
+      
     <a href="https://activepieces.com/docs/pieces/apps/twilio">
       <Card>
           <p align="center">
