+++ docs/pieces/apps/asana.mdx
@@ -14,7 +14,7 @@ title: 'Asana'
 
     <CardGroup cols={2}>
       <Card title="Create Task">
-        Create a new task in a ClickUp workspace and list
+        Create a new task in Asana
       </Card>
     </CardGroup>
