+++ docs/pieces/apps/slack.mdx
@@ -4,8 +4,19 @@ title: 'Slack'
 
 ## Triggers
 
-*No supported triggers yet, please let us know if you need something on Discord so we can help out* 
 
+    <CardGroup cols={2}>
+      <Card title="New Message (Beta)">
+        Trigger when a new message is received
+      </Card>
+    </CardGroup>
+  
+    <CardGroup cols={2}>
+      <Card title="New Reaction (Beta)">
+        Trigger when a new reaction is added to a message
+      </Card>
+    </CardGroup>
+  
 
 ---
