+++ docs/pieces/apps/trello.mdx
@@ -1,0 +1,26 @@
+---
+title: 'trello'
+---
+
+## Triggers
+
+*No supported triggers yet, please let us know if you need something on Discord so we can help out* 
+
+
+---
+
+## Actions
+
+
+    <CardGroup cols={2}>
+      <Card title="Create Card">
+        Create a new card in Trello
+      </Card>
+    </CardGroup>
+  
+    <CardGroup cols={2}>
+      <Card title="Get Card">
+        Get a card in Trello
+      </Card>
+    </CardGroup>
+
