+++ packages/pieces/apps/src/index.ts
@@ -30,7 +30,6 @@ import { posthog } from '@activepieces/piece-posthog';
 import { rssFeed } from '@activepieces/piece-rss';
 import { sendgrid } from '@activepieces/piece-sendgrid';
 import { slack } from '@activepieces/piece-slack';
-import { trello } from './lib/trello';
 import { storage } from '@activepieces/piece-store';
 import { stripe } from '@activepieces/piece-stripe';
 import { telegramBot } from '@activepieces/piece-telegram-bot';
@@ -42,6 +41,7 @@ import { zoom } from '@activepieces/piece-zoom';
 import { generatebanners } from '@activepieces/piece-generatebanners';
 import { connections } from '@activepieces/piece-connections';
 import { youtube } from '@activepieces/piece-youtube';
+import { trello } from '@activepieces/piece-trello';
 
 export const pieces: Piece[] = [
     airtable,
