+++ packages/pieces/trello/README.md
@@ -1,0 +1,7 @@
+# pieces-trello
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running lint
+
+Run `nx lint pieces-trello` to execute the lint via [ESLint](https://eslint.org/).
