+++ packages/pieces/trello/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@activepieces/piece-trello",
+  "version": "0.0.1"
+}
