+++ packages/pieces/trello/src/trello.mdx
@@ -61,3 +61,13 @@ title: 'Trello'
 16. Now click publish, and thats it!
 
 ---
+
+## Triggers
+
+TRIGGERS
+
+---
+
+## Actions
+
+ACTIONS
