+++ packages/pieces/youtube/package.json
@@ -1,4 +1,4 @@
 {
   "name": "@activepieces/piece-youtube",
-  "version": "0.0.1"
+  "version": "0.0.2"
 }
